+++ packages/spotlight/CHANGELOG.md
@@ -10,7 +10,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Changed
 
-- `spotlight/SpotlightContainerDecorator` config property, `enterTo`, default value to be null rather than `'last-focused'`.
+- `spotlight/SpotlightContainerDecorator` config property, `enterTo`, default value to be `null` rather than `'last-focused'`.
 
 ## [1.1.0] - 2017-04-21